--- front_app/src/App.tsx
@@ -1,13 +1,0 @@
-import React from 'react';
-import './App.css';
-import {OSMap} from './components/Map/OSMap';
-
-function App() {
-  return (
-    <div className="App">
-      <OSMap/>
-    </div>
-  );
-}
-
-export default App;
