--- webapp/e2e/jest.config.ts
@@ -1,10 +1,0 @@
-export default {
-    transform: {
-        "^.+\\.tsx?$": "ts-jest"
-    },
-    testMatch: ["**/steps/*.ts"],
-    moduleFileExtensions: ["ts", "tsx", "js", "jsx", "json", "node"],
-    moduleNameMapper:{"^uuid$": "uuid"},
-    preset: "jest-puppeteer",
-    testTimeout: 30000
-}
