--- webapp/jest.config.ts
@@ -1,5 +1,0 @@
-export default {
-    transform: {
-        "^.+\\.tsx?$": "ts-jest"
-    },
-}
