--- webapp/loadtestexample/getuserslist/0004_request.txt
@@ -1,1 +1,0 @@
-{"name":"Pablo","email":"gonzalezgpablo@uniovi.es"}
