+++ webapp/package.json
@@ -1,34 +1,28 @@
 {
-  "name": "webapp",
+  "name": "back_app",
   "version": "0.1.0",
+  "private": true,
   "dependencies": {
-    "@emotion/react": "^11.7.1",
-    "@emotion/styled": "^11.6.0",
-    "@mui/icons-material": "^5.2.5",
-    "@mui/material": "^5.2.5",
-    "@testing-library/jest-dom": "^5.16.1",
-    "@testing-library/react": "^12.1.2",
+    "@testing-library/jest-dom": "^5.16.5",
+    "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^13.5.0",
-    "@types/node": "^18.11.18",
-    "@types/react": "^17.0.37",
-    "@types/react-dom": "^17.0.11",
-    "express": "^4.18.2",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "typescript": "^4.9.4",
-    "web-vitals": "^2.1.2"
-  },
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/Arquisoft/lomap_es6c.git"
+    "@types/jest": "^27.5.2",
+    "@types/node": "^16.18.12",
+    "@types/react": "^18.0.28",
+    "@types/react-dom": "^18.0.11",
+    "leaflet": "^1.9.3",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-leaflet": "^4.2.0",
+    "react-scripts": "5.0.1",
+    "typescript": "^4.9.5",
+    "web-vitals": "^2.1.4"
   },
   "scripts": {
     "start": "react-scripts start",
     "build": "react-scripts build",
-    "test": "react-scripts test --coverage",
-    "test:e2e": "start-server-and-test 'npm --prefix ../restapi start' http://localhost:5000/api/users/list prod 3000 'cd e2e && jest'",
-    "eject": "react-scripts eject",
-    "prod": "ts-node-dev ./server.ts"
+    "test": "react-scripts test",
+    "eject": "react-scripts eject"
   },
   "eslintConfig": {
     "extends": [
@@ -49,20 +43,6 @@
     ]
   },
   "devDependencies": {
-    "@types/expect-puppeteer": "^4.4.7",
-    "@types/express": "^4.17.15",
-    "@types/jest": "^29.2.5",
-    "@types/puppeteer": "^5.4.4",
-    "expect-puppeteer": "^6.0.2",
-    "jest-cucumber": "^3.0.1",
-    "jest-puppeteer": "^6.0.3",
-    "puppeteer": "^13.0.1",
-    "react-scripts": "5.0.0",
-    "start-server-and-test": "^1.14.0",
-    "ts-jest": "^29.0.4",
-    "ts-node-dev": "^2.0.0"
-  },
-  "bugs": {
-    "url": "https://github.com/Arquisoft/lomap_es6c/issues"
+    "@types/leaflet": "^1.9.1"
   }
 }
