+++ webapp/src/App.tsx
@@ -1,38 +1,12 @@
-import React, { useState, useEffect} from 'react';
-import Box from '@mui/material/Box';
-import Link from '@mui/material/Link';
-import Container from '@mui/material/Container';
-import EmailForm from './components/EmailForm';
-import Welcome from './components/Welcome';
-import UserList from './components/UserList';
-import  {getUsers} from './api/api';
-import {User} from './shared/shareddtypes';
+import React from 'react';
 import './App.css';
-import {OSMap} from "./components/Map/OSMap";
+import {OSMap} from './components/Map/OSMap';
 
-function App(): JSX.Element {
-
-  const [users,setUsers] = useState<User[]>([]);
-
-  const refreshUserList = async () => {
-    setUsers(await getUsers());
-  }
-
-  useEffect(()=>{
-    refreshUserList();
-  },[]);
-
+function App() {
   return (
-    <>
-      <Container maxWidth="sm">
-        <OSMap/>
-        <Welcome message="ASW students"/>
-        <Box component="div" sx={{ py: 2}}>This is a basic example of a React application using Typescript. You can add your email to the list filling the form below.</Box>
-        <EmailForm OnUserListChange={refreshUserList}/>        
-        <UserList users={users}/>
-        <Link href="https://github.com/arquisoft/lomap_0">Source code</Link>
-      </Container>
-    </>
+    <div className="App">
+      <OSMap/>
+    </div>
   );
 }
