--- webapp/src/shared/shareddtypes.ts
@@ -1,4 +1,0 @@
-export type User = {
-    name:string;
-    email:string;
-  }
