+++ webapp/tsconfig.json
@@ -13,7 +13,7 @@
     "strict": true,
     "forceConsistentCasingInFileNames": true,
     "noFallthroughCasesInSwitch": true,
-    "module": "commonjs",
+    "module": "esnext",
     "moduleResolution": "node",
     "resolveJsonModule": true,
     "isolatedModules": true,
@@ -21,6 +21,6 @@
     "jsx": "react-jsx"
   },
   "include": [
-    "src","e2e"
+    "src"
   ]
 }
